refactor(extension): merge micro-modules (simplification pass 2)
Structure-only: fewer files, same product behavior.

- github: fold ignoredPrPaths + prFileDiffUrl into prUrls
- review: fold pathClass + fileReviewPlan into reviewPlan
- prefs: auto-open + diff view mode → lib/preferences
- confirmation: move out of overlay/ so options does not import content
- overlay: displayUnits → store; BinaryElided → DiffPane; CloseButton → ModalShell; hunk styles/highlight/extras → hunkShared
- options: merge options.scss into styles.css

Verification: typecheck, 594 unit tests, lint, build:extension.
