+++ apps/extension/src/content/overlay/components/diff/UnifiedHunk.tsx
@@ -1,13 +1,14 @@
 import { useMemo } from "react";
 import { cn } from "@guided-review/ui";
-import { CodeContent, highlightHunkLines } from "./hunkHighlight";
-import { LineExtras } from "./LineExtras";
 import {
+  CodeContent,
   DIFF_LINE_WRAP,
+  highlightHunkLines,
+  LineExtras,
   lineNumberClasses,
   selectionClasses,
   type HunkViewProps,
-} from "./hunkStyles";
+} from "./hunkShared";
 import { lineIdFor, sideForLine } from "../../commentTypes";
 
 export function UnifiedHunk({
