+++ apps/extension/src/content/overlay/components/diff/useSelectionDerived.ts
@@ -6,7 +6,7 @@ import {
   type LineSelection,
   type SelectableLine,
 } from "../../commentTypes";
-import type { ComposerRange } from "./hunkStyles";
+import type { ComposerRange } from "./hunkShared";
 
 export function useSelectionDerived(
   selectableLines: SelectableLine[],
