+++ apps/extension/src/lib/github/prUrls.test.ts
@@ -1,5 +1,11 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { isPrFilesChangedPath, navigateToPrConversation, prConversationUrl } from "./prUrls";
+import {
+  buildPRFileDiffUrl,
+  isIgnoredPrPath,
+  isPrFilesChangedPath,
+  navigateToPrConversation,
+  prConversationUrl,
+} from "./prUrls";
 
 describe("prConversationUrl", () => {
   it("builds the conversation URL", () => {
@@ -32,6 +38,52 @@ describe("isPrFilesChangedPath", () => {
   it("rejects paths that only contain files/changes as a prefix of another segment", () => {
     expect(isPrFilesChangedPath("/acme/widgets/pull/1/filesx")).toBe(false);
     expect(isPrFilesChangedPath("/acme/widgets/pull/1/changesx")).toBe(false);
+  });
+});
+
+describe("isIgnoredPrPath", () => {
+  it("matches the conflicts resolution path", () => {
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/conflicts")).toBe(true);
+    expect(isIgnoredPrPath("/acme/widgets/pull/42/conflicts")).toBe(true);
+  });
+
+  it("matches trailing slash and nested segments under conflicts", () => {
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/conflicts/")).toBe(true);
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/conflicts/file.ts")).toBe(true);
+  });
+
+  it("does not match conversation or other PR tabs", () => {
+    expect(isIgnoredPrPath("/acme/widgets/pull/1")).toBe(false);
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/")).toBe(false);
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/files")).toBe(false);
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/changes")).toBe(false);
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/commits")).toBe(false);
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/checks")).toBe(false);
+  });
+
+  it("does not match lookalike path segments", () => {
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/conflictsx")).toBe(false);
+    expect(isIgnoredPrPath("/acme/widgets/pull/1/xconflicts")).toBe(false);
+  });
+
+  it("does not match non-PR paths", () => {
+    expect(isIgnoredPrPath("/acme/widgets/issues/1/conflicts")).toBe(false);
+    expect(isIgnoredPrPath("/conflicts")).toBe(false);
+    expect(isIgnoredPrPath("/")).toBe(false);
+  });
+});
+
+describe("buildPRFileDiffUrl", () => {
+  // Hash matches GitHub's documented `#diff-` fragment for this path:
+  // https://github.com/orgs/community/discussions/43908
+  it("builds a Files-changed deep link with the path hash", async () => {
+    const url = await buildPRFileDiffUrl(
+      { owner: "acme", repo: "widgets", number: 42 },
+      "src/index.js",
+    );
+    expect(url).toBe(
+      "https://github.com/acme/widgets/pull/42/files#diff-bfe9874d239014961b1ae4e89875a6155667db834a410aaaa2ebe3cf89820556",
+    );
   });
 });
