+++ apps/extension/src/options/main.tsx
@@ -2,7 +2,6 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { App } from "./App";
 import "./styles.css";
-import "./options.scss";
 
 const root = document.getElementById("root");
 if (!root) throw new Error("Options page is missing its #root element.");
